requirements.txt
Pillow==10.0.0
pytest==7.4.0
pytest-cov==4.1.0
black==23.7.0
ruff==0.1.0